requirements.txt
aiomysql==0.0.20
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
openpyxl==3.0.2
peewee==2.10.2
peewee-async==0.5.12
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.2
python-dotenv==0.10.3
redis==3.3.11
requests==2.22.0
six==1.14.0
tornado==6.0.3
urllib3==1.25.7

